refactor: move class "PartidoPolitico" from root package to "models"
