+++ src/main/java/com/evaluacion/votos/evaluacion_backend_votos/EvaluacionBackendVotosApplication.java
@@ -1,0 +1,13 @@
+package com.evaluacion.votos.evaluacion_backend_votos;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class EvaluacionBackendVotosApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(EvaluacionBackendVotosApplication.class, args);
+	}
+
+}
