+++ src/main/resources/application.properties
@@ -1,0 +1,5 @@
+spring.application.name=evaluacion-backend-votos
+spring.datasource.driver-class-name=org.h2.Driver
+spring.datasource.username=sa
+spring.datasource.password=
+spring.datasource.url=jdbc:h2:mem:testdb
