+++ src/test/java/com/evaluacion/votos/evaluacion_backend_votos/EvaluacionBackendVotosApplicationTests.java
@@ -1,0 +1,14 @@
+package com.evaluacion.votos.evaluacion_backend_votos;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class EvaluacionBackendVotosApplicationTests {
+
+	@Test
+	void contextLoads() {
+        throw new UnsupportedOperationException("Not supported yet.");
+	}
+
+}
